Move witness to dedicated file
